files/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
langchain>=0.2.14
langchain-groq>=0.1.9
langgraph>=0.2.0
langchain-core>=0.2.27
groq>=0.9.0
python-multipart>=0.0.9
httpx>=0.27.0
